+++ mono/test/commands/format_command_test.dart
@@ -35,6 +35,7 @@ void main() {
           groupStoreFactory: groupStoreFactory,
           envBuilder: envBuilder,
           plugins: PluginRegistry({}),
+          executor: const DefaultTaskExecutor(),
         );
         expect(code, 1);
         expect(
@@ -77,6 +78,7 @@ void main() {
           groupStoreFactory: groupStoreFactory,
           envBuilder: envBuilder,
           plugins: PluginRegistry({}),
+          executor: const DefaultTaskExecutor(),
         );
         expect(code, 0);
         expect(outCap.text,
@@ -121,6 +123,7 @@ void main() {
           groupStoreFactory: groupStoreFactory,
           envBuilder: envBuilder,
           plugins: PluginRegistry({}),
+          executor: const DefaultTaskExecutor(),
         );
         expect(code, 0);
         expect(
@@ -167,6 +170,7 @@ void main() {
           groupStoreFactory: groupStoreFactory,
           envBuilder: envBuilder,
           plugins: PluginRegistry({}),
+          executor: const DefaultTaskExecutor(),
         );
         expect(code, 0);
         expect(outCap.text, contains('in input order.'));